requirements.txt
python-osc
gbulb
pycairo
hidapi
